+++ src/render_engine/vulkan/vulkan_swapchain.hpp
@@ -35,7 +35,7 @@ namespace nova::renderer::rhi {
                         const std::vector<VkPresentModeKHR>& present_modes);
 
 #pragma region Swapchain implementation
-        uint32_t acquire_next_swapchain_image() override;
+        uint8_t acquire_next_swapchain_image() override;
 
         void present(uint32_t image_idx) override;
 #pragma endregion
